+++ packages/svelte/tests/compiler-errors/samples/runes-no-derived-update/main.svelte
@@ -1,5 +1,5 @@
 <script>
 	const a = $state(0);
-	const b = $derived({ a });
-	b.a++;
+	let b = $derived(a);
+	b++;
 </script>