+++ packages/svelte/tests/runtime-runes/samples/attribute-parts/_config.js
@@ -1,0 +1,6 @@
+import { flushSync } from 'svelte';
+import { test } from '../../test';
+
+export default test({
+	html: `<div class="123"></div><img src="12 hello, world 13">`
+});