+++ packages/svelte/tests/runtime-runes/samples/attribute-parts/main.svelte
@@ -1,0 +1,9 @@
+<script>
+	let a = 1;
+	let b = 2;
+	let c = 3;
+</script>
+
+<div class="{a}{b}{c}" />
+
+<img src="{a}{b} hello, world {a}{c}" />